// api-sankhya-center/src/sankhya/tgfpro/models/tgfpro.interface.ts
export interface Tgfpro {
  codprod: number
  descrprod: string
  compldesc?: string
  caracteristicas?: string
  referencia?: string
  codgrupoprod: number
  codvol: string
  marca?: string
  localizacao?: string
  codipi?: number
  classubtrib?: number
  codformprec?: number
  marclucro?: number
  multipvenda?: number
  decqtd?: number
  decvlr?: number
  deccusto?: number
  comger?: number
  comvend?: number
  descmax?: number
  pesobruto: number
  pesoliq: number
  medaux?: number
  prazoval?: number
  agrupmin?: number
  qtdemb?: number
  alertaestmin: string
  promocao: string
  temicms: string
  temiss: string
  temipivenda: string
  temipicompra: string
  temirf: string
  percirf?: number
  teminss: string
  percinss: number
  redbaseinss: number
  usoprod: string
  origprod: string
  tipsubst: string
  codicmsfast?: number
  tiplancnota: string
  tipcontest: string
  codtab?: number
  codctactb?: number
  imagem?: Buffer
  ativo: string
  estmax?: number
  estmin?: number
  apresdetalhe: string
  selecionado?: string
  titcontest?: string
  liscontest?: string[]
  grupoicms?: number
  percaumentcusto?: number
  local: string
  codmoeda: number
  dtalter: Date
  usalocal: string
  homepage?: string
  codctactb2?: number
  codctactb3?: number
  temposerv?: Date
  endimagem?: string
  codusu: number
  codparcforn?: number
  codvolcompra: string
  codprodroi?: string
  grupodescprod?: string
  refforn?: string
  hrdoobrada: string
  icmsgerencia: string
  codnat: number
  codcencus: number
  codproj: number
  m3?: number
  temciap: string
  implaudolote: string
  codgar?: number
  codctactb4?: number
  dimensoestipo: string
  percquebratec?: number
  codfiltro?: number
  endmodrotulo?: string
  codgenero?: number
  codlst?: number
  padrao: string
  naturezaoperdes?: string
  cnae?: number
  solcompra: string
  confere: string
  remeter: string
  motivoincexc?: string
  arredpreco?: number
  temcomissao: string
  componobrigatorio?: string
  fattotal?: string
  nometab?: string
  ap1rctego: string
  calculogiro: string
  redbaseirf?: number
  altura?: number
  largura?: number
  espessura?: number
  unidade: string
  codformaponta?: number
  codcos?: number
  confcegapeso: string
  ordcalccustprod?: number
  regrawms: string
  grupopis: string
  grupocofins: string
  grupocssl: string
  cstipisent?: number
  cstipisai?: number
  statusincexc?: string
  utilizawms: string
  balanca: string
  codpais?: number
  rendimento?: string
  obsetiqueta?: string
  codanp?: number
  codauteccodif?: string
  codprodagrupapt?: number
  codprodagrupaptenc?: number
  cultura?: string
  cientifico?: string
  classeagt?: number
  grupoquimico?: number
  classtox?: number
  prinativo?: number
  formulacao?: string
  concentracao?: string
  modoaplic?: string
  epocaaplic?: string
  manejoint?: string
  dosagem?: number
  voldosagem?: string
  dosagempor?: number
  voldosagempor?: string
  rendimentopr?: string
  receituario: string
  exigbenefic: string
  tipoclasseagt?: string
  tipogrupoquimico?: string
  tipoprinativo?: string
  tipoclasstox?: string
  grupotransg?: number
  geraplaprod: string
  aplicacao?: number
  intervalo?: number
  carencia?: number
  rendimentoha?: number
  produtonfe: number
  tipgtinnfe: number
  ncm?: string

  estoque?: {
    estoqueTotal: number
    estminTotal: number
    estmaxTotal: number
    locais: number
    abaixoMinimo: boolean
    locaisNomes?: string[]
  }

  tgfgru?: { codgrupoprod: number; descgrupoprod: string }
  tgfvol?: { codvol: string; descrvol: string }
  tgfpar?: { codparc: number; nomeparc: string }
  tgfnat?: { codnat: number; descnat: string }
  tgfcencus?: { codcencus: number; descricao: string }
  tgfproj?: { codproj: number; identproj: string }
}